requirements.txt
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rapidfuzz>=3.0.0
pypdf>=3.0.0
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
transformers>=4.40.0
torch>=2.0.0; sys_platform != "linux"
sentencepiece>=0.2.0
sacremoses>=0.1.0
